Update Python for FPGA
This Python Code is the continuation of the HDL project written in the VHDL language. The main goal of this code is to communicate with the FPGA, sending random inputs, holding the outputs transmitted from the UART of FPGA (results of the ALU machine), finding expected results, comparing them with the outputs of the ALU machine, and finally print the results on a PDF file.